split up into modules and add docs
