+++ crates/float/Cargo.toml
@@ -29,5 +29,6 @@ getrandom = { version = "0.2.11", features = ["js", "js-sys"] }
 
 [dev-dependencies]
 alloy = { workspace = true, features = ["sol-types", "json-rpc", "arbitrary"] }
+anyhow = "1.0.98"
 proptest.workspace = true
 serde_json = "1.0.140"
